+++ src/info.ml
@@ -167,6 +167,8 @@ let to_string_mach t = Sexp.to_string_mach (sexp_of_t t)
 
 let of_lazy l = lazy (protect (fun () -> String (Lazy.force l)))
 
+let of_lazy_t lazy_t = Lazy.join lazy_t
+
 let of_string message = Lazy.from_val (String message)
 
 let createf format = Printf.ksprintf of_string format