+++ src/hash_set_intf.ml
@@ -190,6 +190,9 @@ module type Hash_set = sig
   end
 
   (**/**)
+  (*_ See the Jane Street Style Guide for an explanation of [Private] submodules:
+
+    https://opensource.janestreet.com/standards/#private-submodules *)
   module Private : sig
     val hashable : 'a t -> 'a Hashable.t
   end