+++ CHANGES.md
@@ -12,6 +12,9 @@
 
 - Added `Info.of_lazy_t`
 
+- Added `List.partition_result` function, to partition a list of `Result.t`
+  values
+
 ## v0.11
 
 - Deprecated `Not_found`, people who need it can use `Caml.Not_found`, but its